+++ common/reactive/src/main/java/is/codion/common/reactive/value/ObservableValueCollection.java
@@ -51,6 +51,20 @@ public interface ObservableValueCollection<T, C extends Collection<T>> extends O
 	boolean containsAll(Collection<T> values);
 
 	/**
+	 * Returns true if this {@link ValueCollection} instance contains only the elements of the specified collection
+	 * @param values the elements to check
+	 * @return true if this {@link ValueCollection} instance contains only the elements of the specified collection
+	 */
+	boolean containsOnly(Collection<T> values);
+
+	/**
+	 * Returns true if this {@link ValueCollection} instance contains none of the elements of the specified collection
+	 * @param values the elements to check
+	 * @return true if this {@link ValueCollection} instance contains none of the elements of the specified collection
+	 */
+	boolean containsNone(Collection<T> values);
+
+	/**
 	 * @return true if this {@link ValueCollection} instance is empty
 	 */
 	boolean isEmpty();